--- src/const.js
@@ -1,20 +1,0 @@
-export const FilterType = {
-  all: `All movies`,
-  watchlist: `Watchlist`,
-  history: `History`,
-  favorites: `Favorites`
-};
-
-export const GENRES = [
-  `Musical`,
-  `Western`,
-  `Drama`,
-  `Comedy`,
-  `Cartoon`,
-  `Mystery`,
-  `Film-Noir`,
-  `Action`,
-  `Horror`,
-  `Animation`,
-  `Adventure`,
-];
