+++ src/util/const.js
@@ -1,0 +1,15 @@
+export const FilterType = {
+  all: `All movies`,
+  watchlist: `Watchlist`,
+  history: `History`,
+  favorites: `Favorites`
+};
+
+export const Key = {
+  ESCAPE: `Escape`,
+};
+
+export const getRandomArrayItem = (array) => {
+  const randomIndex = Math.floor(Math.random() * array.length);
+  return array[randomIndex];
+};
